backend/requirements.txt
anyio==3.4.0
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
fastapi==0.70.1
fastapi_msal==0.1.7
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
msal==1.16.0
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
python-dotenv==0.19.2
python-multipart==0.0.5
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
